backend/requirements.txt
fastapi>=0.111.0
uvicorn[standard]>=0.29.0
# ollama  # local inference fallback — requires 10GB+ RAM
google-genai
langchain-community==0.4.1
langchain-ollama==1.1.0
langchain-text-splitters==1.1.2
chromadb==1.5.9
pypdf>=4.0.0
python-docx>=1.1.0
python-multipart>=0.0.9
pydantic>=2.7.0
pydantic-settings>=2.2.0
aiofiles>=23.2.0
pillow>=10.0.0
reportlab>=4.2.0
python-dotenv>=1.0.0
httpx>=0.27.0
